UI List: renaming can now also be done with double click and ctrl+enter.
This also fixes a bug where ctrl+click would not properly work when pressing
ctrl after moving the mouse over the text, it had to be pressed beforehand.

Hack is now to have LISTROW active, but when it detects any of these events
it will activate the text field that the mouse overlaps with instead.

Reviewed By: mont29

Differential Revision: http://developer.blender.org/D38